requirements.txt
Django==5.2.4
djangorestframework==3.15.1
Pillow==11.3.0
requests==2.32.4
paypalrestsdk==1.13.3